// campus-wayfinder-ui/src/app/models/event.model.ts
import { Location } from './location.model';

export interface Event {
  id?: number; // Optional, as it may be generated by the database
  eventName: string; // Title of the event
  description?: string; // Description of the event
  startTime?: Date; // Start date and time of the event
  endTime?: Date; // End date and time of the event
  location: Location; // Location of the event, storing full location details
}